finish project

+++ quanlythuvien/QuanLySach.cs
@@ -1,4 +1,5 @@
-﻿using quanlythuvien.Models;
+﻿using quanlythuvien.Infrastructure;
+using quanlythuvien.Models;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -45,7 +46,12 @@ namespace quanlythuvien
 
         private void btnAddBook_Click(object sender, EventArgs e)
         {
-            AddBook addBook = new AddBook();
+            if(_employeeModel.Role != Const.ROLEADMIN)
+            {
+                MessageBox.Show("Bạn Không Có Quyền Thêm Sách");
+                return;
+            }
+            AddBook addBook = new AddBook(_employeeModel);
             addBook.Show();
             this.Hide();
 
@@ -53,7 +59,13 @@ namespace quanlythuvien
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
+            if (_employeeModel.Role != Const.ROLEADMIN)
+            {
+                MessageBox.Show("Bạn Không Có Quyền Xóa Sách");
+                return;
+            }
             DeleteBook(Convert.ToInt32(tbId.Text));
+            MessageBox.Show("Xóa Thành Công");
         }
 
         void GetAllBooks()
@@ -67,6 +79,11 @@ namespace quanlythuvien
 
         private void btnUpdateBook_Click(object sender, EventArgs e)
         {
+            if (_employeeModel.Role != Const.ROLEADMIN)
+            {
+                MessageBox.Show("Bạn Không Có Quyền Sửa Sách");
+                return;
+            }
             BookModel bookModel = new BookModel();
             bookModel = GetBookById(Convert.ToInt32(tbId.Text));
             UpdateBook updateBook = new UpdateBook(_employeeModel, bookModel);
@@ -119,36 +136,6 @@ namespace quanlythuvien
             }
             return bookModel;
 
-        }
-
-        BookCategoryModel GetCategoryById(int id)
-        {
-            BookCategoryModel bookCategoryModel = null;
-            try
-            {
-                using (SqlConnection connection = new SqlConnection(connectionString))
-                {
-                    connection.Open();
-
-                    using (SqlCommand command = new SqlCommand("GetCategoryById", connection))
-                    {
-                        command.CommandType = CommandType.StoredProcedure;
-                        command.Parameters.Add(new SqlParameter("@Id", SqlDbType.Int)).Value = id;
-                        using (SqlDataReader reader = command.ExecuteReader())
-                        {
-                            if (reader.Read())
-                            {
-                                bookCategoryModel = new BookCategoryModel
-                                { CategoryName = reader["TenTheLoai"].ToString() };
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-            }
-            return bookCategoryModel;
         }
 
         PublisherModel GetPublisherById(int id)
@@ -208,6 +195,11 @@ namespace quanlythuvien
 
         private void btnNhapHang_Click(object sender, EventArgs e)
         {
+            if (_employeeModel.Role != Const.ROLEADMIN)
+            {
+                MessageBox.Show("Bạn Không Có Quyền Xem Hóa Đơn");
+                return;
+            }
             HoaDonNhapHang hdnh = new HoaDonNhapHang(_employeeModel);
             hdnh.Show();
             this.Hide();
@@ -218,9 +210,9 @@ namespace quanlythuvien
             AddToCard();
         }
 
-        int GetEmployeeByName(string name)
+        EmployeeModel GetEmployeeByName(string name)
         {
-            int id = 0;
+            EmployeeModel employeeModel = null;
             name = tbCustonmerName.Text;
             try
             {
@@ -232,10 +224,18 @@ namespace quanlythuvien
                     {
                         command.CommandType = CommandType.StoredProcedure;
                         command.Parameters.Add(new SqlParameter("@Name", SqlDbType.NVarChar)).Value = name;
-                        object result = command.ExecuteScalar();
-                        if (result != null && result != DBNull.Value)
+                        using (SqlDataReader reader = command.ExecuteReader())
                         {
-                            id = Convert.ToInt32(result);
+                            while (reader.Read())
+                            {
+                                employeeModel = new EmployeeModel
+                                {
+                                    Id = Convert.ToInt32(reader["Id"]),
+                                    EmployeeName = reader["HoTen"].ToString(),
+                                    Role = reader["TenRole"].ToString()
+                                    
+                                };
+                            }
                         }
                     }
                 }
@@ -243,7 +243,7 @@ namespace quanlythuvien
             catch (Exception ex)
             {
             }
-            return id;
+            return employeeModel;
 
         }
 
@@ -265,9 +265,21 @@ namespace quanlythuvien
                         command.ExecuteNonQuery();
                         MessageBox.Show("Thêm Giỏ Hàng Thành Công");
                     }
-                    catch (Exception ex)
+                    catch (SqlException ex)
                     {
+                        if (ex.Errors.Count > 0)
+                        {
+                            foreach (SqlError error in ex.Errors)
+                            {
+                                if (error.Number == 50000) 
+                                {
+                                    MessageBox.Show(error.Message);
+                                    return;
+                                }
+                            }
+                        }
                     }
+
                 }
             }
             
@@ -282,8 +294,47 @@ namespace quanlythuvien
 
         private void btnbanhang_Click(object sender, EventArgs e)
         {
+            if (_employeeModel.Role != Const.ROLEADMIN)
+            {
+                MessageBox.Show("Bạn Không Có Quyền Xem Hóa Đơn");
+                return;
+            }
             HoaDonBanHang hd = new HoaDonBanHang(_employeeModel);
             hd.Show();
+            this.Hide();
+        }
+
+        public void SearchBooks(string searchTerm)
+        {
+
+
+            using (SqlConnection connection = new SqlConnection(connectionString))
+            {
+                connection.Open();
+
+                using (SqlCommand command = new SqlCommand("SearchBooks", connection))
+                {
+                    dataGridView1.DataSource = null;
+                    command.CommandType = CommandType.StoredProcedure;
+                    command.Parameters.AddWithValue("@SearchTerm", searchTerm);
+                    SqlDataAdapter adapter = new SqlDataAdapter(command);
+                    DataTable dt = new DataTable();
+                    adapter.Fill(dt);
+                    dataGridView1.DataSource = dt;
+
+                }
+            }
+
+        }
+        private void btnSearch_Click(object sender, EventArgs e)
+        {
+            SearchBooks(tbSearch.Text);
+        }
+
+        private void btnDangXuat_Click(object sender, EventArgs e)
+        {
+            Login lg = new Login();
+            lg.Show();
             this.Hide();
         }
     }
